flash430: Store the firmware version in the MSP430 flash.
This patch makes flash430, the I2C bootloader, store a 2-byte firmware
revision number at the beginning of flash.  It also connects up the
I2C handler to this value.

The value is stored in flash using by adding a new section, ".fwver",
which the linker script places at the beginning of flash.

git-svn-id: http://svn.srobo.org/boards/lib-fw/flash430/trunk@2271 98f67ba8-b320-0410-bda0-b753e294843c