Disable WDT while modifying flash
The WDT on the power board has a period of 4ms (it can't be made
longer than this unfortunately). Erasing and writing to the flash
takes longer than this so the WDT must be disabled while doing so.
The WDT is only re-enabled if it was enabled before modifying the
flash.